Renamespace ruleset masks
