+++ osu.Game.Rulesets.Mania/Edit/Masks/HoldNoteSelectionMask.cs
@@ -13,7 +13,7 @@ using osu.Game.Rulesets.UI.Scrolling;
 using OpenTK;
 using OpenTK.Graphics;
 
-namespace osu.Game.Rulesets.Mania.Edit.Layers.Selection.Overlays
+namespace osu.Game.Rulesets.Mania.Edit.Masks
 {
     public class HoldNoteSelectionMask : SelectionMask
     {
